add example of ideal flow with explicitly creating an auth token
generally this would work, but we run into the problem of needing to
pass the created client object around to each promise. this doesn't
seem scalable.

it would be better if we could attach the created client back to the
exported module, but only allow it's usage if it has been created.